refactor(simulation): separate preparation, execution and result receipts
Extract pure input preparation, revision-scoped identity cache, execution port, bounded receipts, deck/metadata/verdict modules. Publish full output artifacts and preserve cancellation retry semantics. Reuse canonical environment and path validation. Existing run registry remains the sole lifecycle owner.

Validation: 230 focused simulation, worker, local-host and UI tests pass; typecheck passes. Gate plan requires full delivery due to shared contracts; full branch gate follows the UI integration commit. Other dirty files are subsequent targets in this refactor.

Test-Impact: tests-updated
