+++ packages/simulation-service/src/index.ts
@@ -5,3 +5,5 @@ export * from "./hosted-executor.js";
 export * from "./result-volume.js";
 export * from "./output-evaluation.js";
 export * from "./automatic-measurements.js";
+
+export * from "./executor.js";
